Refactor distro-specific code out of App
