+++ src/App.jsx
@@ -12,10 +12,10 @@ import { Provider, useSelector } from 'react-redux'
 import { PersistGate } from 'redux-persist/integration/react'
 import { PaperProvider } from 'react-native-paper'
 import MaterialCommunityIcon from 'react-native-vector-icons/MaterialCommunityIcons'
-import codePush from 'react-native-code-push'
-import { Provider as RollbarProvider, ErrorBoundary } from '@rollbar/react'
 
 import 'react-native-gesture-handler' // This must be included in the top component file
+
+import GLOBAL from './GLOBAL'
 
 import { usePrepareThemes } from './styles/tools/usePrepareThemes'
 
@@ -30,33 +30,7 @@ import OperationAddActivityScreen from './screens/OperationScreens/OperationAddA
 import OperationActivityOptionsScreen from './screens/OperationScreens/OperationActivityOptionsScreen'
 import OperationBadgeScreen from './screens/OperationBadgeScreen/OperationBadgeScreen'
 import MainSettingsScreen from './screens/SettingsScreens/screens/MainSettingsScreen'
-
-const GLOBAL_APP_SETTINGS = {
-  consentAppData: false
-}
-
-const DISTRIBUTION_CONFIG = {}
-
-/** EXAMPLE CODEPUSH CONFIG */
-// DISTRIBUTION_CONFIG.codePushOptions = {
-//   installMode: codePush.InstallMode.IMMEDIATE
-// }
-
-/** EXAMPLE ROLLBAR CONFIG */
-// import { Client } from 'rollbar-react-native'
-// DISTRIBUTION_CONFIG.rollbarNative = new Client({
-//   accessToken: Config.ROLLBAR_TOKEN,
-//   captureUncaught: true,
-//   captureUnhandledRejections: true
-// })
-
-export function reportError (error, ...extra) {
-  if (GLOBAL_APP_SETTINGS.consentAppData && DISTRIBUTION_CONFIG.rollbarNative && DISTRIBUTION_CONFIG.rollbarNative.rollbar) {
-    DISTRIBUTION_CONFIG.rollbarNative.rollbar.error(error, ...extra)
-  }
-  console.error(error, ...extra)
-  if (extra && extra[0]?.stack) console.error(extra[0].stack)
-}
+import { AppWrappedForDistribution, useConfigForDistribution } from './distro'
 
 const Stack = createNativeStackNavigator()
 
@@ -68,8 +42,11 @@ function MainApp ({ navigationTheme }) {
   const [appState, setAppState] = useState('starting')
 
   const settings = useSelector(selectSettings)
+
+  useConfigForDistribution({ settings })
+
   useEffect(() => { // Some top-level functions need access to settings info that's only available in the store at this point
-    GLOBAL_APP_SETTINGS.consentAppData = settings.consentAppData
+    GLOBAL.consentAppData = settings.consentAppData
   }, [settings?.consentAppData])
 
   if (appState === 'starting') {
@@ -129,29 +106,14 @@ function ThemedApp () {
   )
 }
 
-const PersistedApp = () => (
-  <Provider store={store}>
-    <PersistGate loading={null} persistor={persistor}>
-      <ThemedApp />
-    </PersistGate>
-  </Provider>
+const App = () => (
+  <AppWrappedForDistribution>
+    <Provider store={store}>
+      <PersistGate loading={null} persistor={persistor}>
+        <ThemedApp />
+      </PersistGate>
+    </Provider>
+  </AppWrappedForDistribution>
 )
-
-let App
-if (DISTRIBUTION_CONFIG.rollbarNative && DISTRIBUTION_CONFIG.rollbarNative.rollbar) {
-  App = () => (
-    <RollbarProvider instance={DISTRIBUTION_CONFIG.rollbarNative.rollbar}>
-      <ErrorBoundary>
-        <PersistedApp />
-      </ErrorBoundary>
-    </RollbarProvider>
-  )
-} else {
-  App = PersistedApp
-}
-
-if (DISTRIBUTION_CONFIG.codePushOptions) {
-  App = codePush(DISTRIBUTION_CONFIG.codePushOptions)(App)
-}
 
 export default App
