+++ src/GLOBAL.js
@@ -1,0 +1,11 @@
+/*
+ * Copyright ©️ 2024 Sebastian Delmont <sd@ham2k.com>
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public License, v. 2.0.
+ * If a copy of the MPL was not distributed with this file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ */
+
+const GLOBAL = {
+  consentAppData: false
+}
+export default GLOBAL
