+++ src/distro/devDistributionConfig.js
@@ -1,0 +1,25 @@
+/*
+ * Copyright ©️ 2024 Sebastian Delmont <sd@ham2k.com>
+ *
+ * This Source Code Form is subject to the terms of the Mozilla Public License, v. 2.0.
+ * If a copy of the MPL was not distributed with this file, You can obtain one at https://mozilla.org/MPL/2.0/.
+ */
+
+import React from 'react'
+
+export function reportError (error, ...extra) {
+  console.error(error, ...extra)
+  if (extra && extra[0]?.stack) console.error(extra[0].stack)
+}
+
+export function AppWrappedForDistribution ({ children }) {
+  return (
+    <>
+      {children}
+    </>
+  )
+}
+
+export function useConfigForDistribution ({ settings }) {
+  // Do nothing
+}
