+++ src/distro/index.js
@@ -1,0 +1,3 @@
+/* This file particular file is dedicated to the public domain */
+
+export * from './devDistributionConfig'
