+++ src/extensions/commands/commandHandling.js
@@ -5,7 +5,8 @@
  * If a copy of the MPL was not distributed with this file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
 
-import { reportError } from '../../App'
+import { reportError } from '../../distro'
+
 import { findHooks } from '../registry'
 
 export function checkAndProcessCommands (value, extraParams) {
