+++ src/screens/OperationScreens/OpSettingsTab/components/LocationDialog.js
@@ -13,7 +13,8 @@ import { locationToGrid } from '@ham2k/lib-maidenhead-grid'
 
 import { setOperationData } from '../../../../store/operations'
 import ThemedTextInput from '../../../components/ThemedTextInput'
-import { reportError } from '../../../../App'
+import { reportError } from '../../../../distro'
+
 import { Ham2kDialog } from '../../../components/Ham2kDialog'
 
 const VALID_MAIDENHEAD_REGEX = /^([A-R]{2}|[A-R]{2}[0-9]{2}|[A-R]{2}[0-9]{2}[a-x]{2}||[A-R]{2}[0-9]{2}[a-x]{2}[0-9]{2})$/
